allow calling server start, accept, stop individually
This lets external code manage server startup and when new
connections are accepted. The case I have in mind is a set of tests
that fork/exec a command that make an HTTP request. I'd like to
start a server in the test process, exec the command, accept a
single connection, and then check that the command generated the
correct output.